+++ glean-core/rlb/src/private/custom_distribution.rs
@@ -57,7 +57,7 @@ impl glean_core::traits::CustomDistribution for CustomDistributionMetric {
         &self,
         ping_name: S,
     ) -> Option<DistributionData> {
-        dispatcher::block_on_queue();
+        crate::block_on_dispatcher();
 
         let queried_ping_name = ping_name
             .into()
@@ -71,7 +71,7 @@ impl glean_core::traits::CustomDistribution for CustomDistributionMetric {
         error: ErrorType,
         ping_name: S,
     ) -> i32 {
-        dispatcher::block_on_queue();
+        crate::block_on_dispatcher();
 
         crate::with_glean_mut(|glean| {
             glean_core::test_get_num_recorded_errors(&glean, self.0.meta(), error, ping_name.into())